added patients entity

+++ src/main/java/med/jsrdev/api/user/User.java
@@ -1,3 +1,8 @@
 package med.jsrdev.api.user;
 
-public record User(int userId, int id, String title, String body) { }
+public record User(
+        Integer userId,
+        Long id,
+        String title,
+        String body
+) { }
